feat：在删除历史数据时调用 confirm

+++ src/views/layout/components/header/header-search/history.vue
@@ -34,6 +34,7 @@ const EMITS_ITEM_CLICK = 'itemClick'
 
 <script setup>
 import { useStore } from 'vuex'
+import { confirm } from '@/libs'
 
 const emits = defineEmits([EMITS_ITEM_CLICK])
 
@@ -43,7 +44,13 @@ const store = useStore()
  * 删除所有记录
  */
 const onDeleteAllClick = () => {
-  store.commit('search/deleteAllHistory')
+  confirm('title', '要删除所有历史记录吗？')
+    .then(() => {
+      store.commit('search/deleteAllHistory')
+    })
+    .catch(() => {
+      console.log('取消删除')
+    })
 }
 
 /**
